Split up the winding number implementation in two
